backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
PyJWT==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
